agent/requirements.txt
-e ../../openai-agents-python
openai>=1.87.0
python-dotenv
